Исправлены баги и уязвимости. Добавлена валидация. Улучшена модерация. В проектах вместо комиссии устанавливается цена за действие.

+++ partner_app/views/partnership.py
@@ -1,4 +1,5 @@
-from django.shortcuts import redirect
+from django.shortcuts import redirect, get_object_or_404
+from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.db import transaction
 from django.views.decorators.http import require_POST
@@ -12,19 +13,26 @@ from django.utils.timezone import now
 def stop_partnership_with_partner(request,partner_id):
     """Остановить сотрудничество рекламодателя с партнёром (Со всеми подключенными проектами)"""
     
+    
     partnership = ProjectPartner.objects.filter(
         advertiser=request.user,
         partner=partner_id
     )
+    
+    if not partnership.exists():
+        messages.error(request, "Сотрудничество с данным партнёром не найдено.",extra_tags="stop_partnership_error")
+        return redirect('dashboard')
+    
     user = User.objects.get(id=partner_id)
     partnership.delete()
     date_str = now().strftime("%d.%m.%Y %H:%M")
     
-    title = ""
+    title = "❌ Остановка сотрудничества"
     message = f"""Здравствуйте,{user.get_full_name()}!\n\n
 Рекламодатель {request.user.get_full_name()} прекратил сотрудничество с вами {date_str}.\n\n\n
 С уважением,\nКоманда поддержки"""
-    send_email_via_mailru(user.email,message,"❌ Остановка сотрудничества")
+    send_email_via_mailru(user.email,message,title)
+    messages.success(request,message="Сотрудничество с партнёром успешно остановлено!",extra_tags="stop_partnership_success")
     return redirect('dashboard')
 
 @login_required
@@ -32,22 +40,22 @@ def stop_partnership_with_partner(request,partner_id):
 @transaction.atomic
 def stop_partnership_with_project(request,project_id):
     """Остановить сотрудничество партнёра с проектом рекламодателя"""
-    partnership = ProjectPartner.objects.get(partner=request.user,project=project_id)
+    partnership = get_object_or_404(ProjectPartner,partner=request.user,project=project_id)
     partnership.delete()
     
     title = '❌ Остановка сотрудничества'
     message = f"""Уважаемый рекламодатель,
-Сообщаем вам, что партнёр {partnership.partner.get_full_name()} {partnership.partner.email} прекратил сотрудничество по проекту «{partnership.project.name}»
-Причина: {request.POST.get('suspension_reason', "Не указана")}
+сообщаем вам, что партнёр {partnership.partner.get_full_name()} {partnership.partner.email} прекратил сотрудничество по проекту «{partnership.project.name}»\n
+Причина: {request.POST.get('suspension_reason', "Не указана")}\n
 Комментарий: {request.POST.get('suspension_comment', "Не указан")}
 **Что это значит:**  
 - Все ссылки партнёра деактивированы  
 - Новые переходы с его ресурсов не учитываются  
-- Статистика доступна в личном кабинете  
-
+- Статистика доступна в личном кабинете\n\n
 Это письмо отправлено автоматически."""    
 
     send_email_via_mailru(partnership.advertiser.email,message,title)
+    messages.success(request,message="Сотрудничество с партнёром успешно остановлено!",extra_tags="stop_partnership_success")
     return redirect('dashboard')
 
 @login_required
@@ -73,6 +81,7 @@ def suspend_partnership(request,project_id):
 Это письмо отправлено автоматически."""
     
     send_email_via_mailru(partnership.advertiser.email,message,title)
+    messages.success(request,message="Сотрудничество с партнёром успешно приостановлено!",extra_tags="suspend_partnership_success")
     return redirect('dashboard')
 
 
@@ -80,7 +89,7 @@ def suspend_partnership(request,project_id):
 @require_POST
 def resume_partnership(request,project_id):
     """ Воозобновить сотрудничество партнёра с проектом рекламодателя"""
-    partnership = ProjectPartner.objects.get(partner=request.user,project=project_id)
+    partnership = get_object_or_404(ProjectPartner,partner=request.user,project=project_id)
     partnership.status = "Активен"
     partnership.suspension_reason = None 
     partnership.suspension_comment = None
@@ -92,5 +101,5 @@ def resume_partnership(request,project_id):
 После возобновления сотрудничества у вас будут учитываться конверсии/переходы.
 Это письмо отправлено автоматически."""
     send_email_via_mailru(partnership.advertiser.email,message,title)
-    
+    messages.success(request,message="Сотрудничество с партнёром успешно возобновлено!",extra_tags="resume_partnership_success")
     return redirect('dashboard')
